+++ compiler/frontend/src/org/jetbrains/kotlin/psi/KtTypeArgumentList.java
@@ -18,6 +18,7 @@ package org.jetbrains.kotlin.psi;
 
 import com.intellij.lang.ASTNode;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.kotlin.lexer.KtTokens;
 import org.jetbrains.kotlin.psi.stubs.KotlinPlaceHolderStub;
 import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes;
 
@@ -40,5 +41,10 @@ public class KtTypeArgumentList extends KtElementImplStub<KotlinPlaceHolderStub<
     @NotNull
     public List<KtTypeProjection> getArguments() {
         return getStubOrPsiChildrenAsList(KtStubElementTypes.TYPE_PROJECTION);
+    }
+
+    @NotNull
+    public KtTypeProjection addArgument(@NotNull KtTypeProjection typeArgument) {
+        return EditCommaSeparatedListHelper.INSTANCE.addItem(this, getArguments(), typeArgument, KtTokens.LT);
     }
 }