+++ idea/testData/quickfix/createFromUsage/createClass/delegationSpecifier/classDelegatorToSuperclass.kt
@@ -1,5 +1,6 @@
 // "Create class 'A'" "false"
 // ACTION: Create interface 'A'
+// ACTION: Create type parameter 'A' in class 'Foo'
 // ACTION: Create test
 // ERROR: Unresolved reference: A
 package p