+++ idea/testData/quickfix/createFromUsage/createTypeParameter/typeQualifier.kt
@@ -1,0 +1,12 @@
+// "Create type parameter 'X'" "false"
+// ACTION: Create class 'X'
+// ACTION: Create enum 'X'
+// ACTION: Create interface 'X'
+// ACTION: Create object 'X'
+// ERROR: Unresolved reference: X
+
+class A
+
+fun foo(x: <caret>X.Y) {
+
+}