+++ idea/testData/quickfix/createFromUsage/createTypeParameter/withTypeArguments.kt
@@ -1,0 +1,10 @@
+// "Create type parameter 'X'" "false"
+// ACTION: Create class 'X'
+// ACTION: Create interface 'X'
+// ERROR: Unresolved reference: X
+
+class A
+
+fun foo(x: <caret>X<Int>) {
+
+}